requirements.txt
python-telegram-bot>=20.3
pymongo>=4.3.3
mongoengine>=0.24.1
dnspython>=2.4.2
configparser>=5.3.0
telegraph>=1.4.1
